docs: switch to BYOCNI in AKS prerequisites / instructions
Following up on the previous commit: we add documentation related to the
new AKS BYOCNI capabilities, and recommend it as the preferred
installation method for AKS clusters.

Since users might still want to use the integration offered by Azure
IPAM, we elect to keep both the older Azure IPAM-based AKS installation
method and the new BYOCNI-based installation method in the
documentation.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>