nodediscovery: split StartDiscovery into smaller methods
[ upstream commit c19ef1c36ba890945507a59e3af642db8c743f27 ]

in preparation to a subsequent change, this commit moves some of the
(*NodeDiscovery).StartDiscovery() logic into separated methods:

* fillLocalNode(), responsible for syncing the localNode object with the
  actual state of the node
* updateLocalNode(), responsible for publishing the updated KV store
  entry and/or CiliumNode object for the local node

In addition to that, this commit introduces also the UpdateLocalNode()
method, which ensures that the proper lock is held while syncing the
localNode object state with fillLocalNode() and then publishing the
new object with updateLocalNode().

No functional changes are introduced.

Signed-off-by: Gilberto Bertin <jibi@cilium.io>