daemon: Switch policy over to new IPCache interface
Convert the daemon over to using the new IPCache interface which handles
merging of metadata from multiple sources. As a side effect, we can
clean up some of the identity refcounting/management handling code from
the daemon in favour of the core code in pkg/ipcache.

This patch may subtly change the ordering of operations for policies
that have {to,from}CIDR{,Set} rules in them. Previously:

1. Figure out the prefixes
2. Synchronously allocate identities for the prefixes
3. Regenerate policy for all relevant endpoints
4. Trigger IPCache updates asynchronously, including
4a. Trigger SelectorCache / policy updates via IPCache
4b. Update datapath IPCache

Step (2) may now occur after 3 but will continue to occur before (4).
This means that step (3) will only prepare the SelectorCache and policy
constructs associated with the endpoints, but not associate the new
identities for the CIDRs to those policy constructs. That latter step is
deferred to the incremental "identity update" logic in (4a).

Signed-off-by: Joe Stringer <joe@cilium.io>